Moving JwtDecoder to a new class
