+++ core/src/main/java/io/aiven/klaw/config/SecurityConfigSSO.java
@@ -1,20 +1,7 @@
 package io.aiven.klaw.config;
 
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.core.type.TypeReference;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.nimbusds.jose.Payload;
-import com.nimbusds.jwt.JWT;
-import com.nimbusds.jwt.JWTClaimsSet;
-import com.nimbusds.jwt.JWTParser;
-import com.nimbusds.jwt.SignedJWT;
 import io.aiven.klaw.auth.KwAuthenticationFailureHandler;
 import io.aiven.klaw.auth.KwAuthenticationSuccessHandler;
-import java.text.ParseException;
-import java.util.Date;
-import java.util.HashMap;
-import java.util.LinkedHashMap;
-import java.util.Map;
 import java.util.Properties;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -35,10 +22,7 @@ import org.springframework.security.oauth2.client.web.HttpSessionOAuth2Authoriza
 import org.springframework.security.oauth2.client.web.OAuth2AuthorizedClientRepository;
 import org.springframework.security.oauth2.client.web.reactive.function.client.ServletOAuth2AuthorizedClientExchangeFilterFunction;
 import org.springframework.security.oauth2.core.endpoint.OAuth2AuthorizationRequest;
-import org.springframework.security.oauth2.jwt.Jwt;
-import org.springframework.security.oauth2.jwt.JwtDecoder;
 import org.springframework.security.oauth2.jwt.JwtDecoderFactory;
-import org.springframework.security.oauth2.jwt.JwtException;
 import org.springframework.security.provisioning.InMemoryUserDetailsManager;
 import org.springframework.security.web.SecurityFilterChain;
 import org.springframework.web.reactive.function.client.WebClient;
@@ -99,96 +83,10 @@ public class SecurityConfigSSO {
     return new InMemoryUserDetailsManager(globalUsers);
   }
 
-  /*
-  Below code and class is only applicable for keycloak provider with legacy version (10.x)
-  Known issue in keycloak https://issues.redhat.com/browse/KEYCLOAK-14309
-  If token contains duplicate keys, json parser in spring framework fails. Hence custom json parser is applied here below.
-  */
-
+  // Below code is only applicable for keycloak provider with legacy version (10.x)
   @Bean
   @ConditionalOnExpression("${klaw.sso.provider.keycloak.legacy:false}")
   public JwtDecoderFactory<ClientRegistration> customJwtDecoderFactory() {
-    return new CustomJwtDecoderFactory();
-  }
-
-  static class CustomJwtDecoderFactory implements JwtDecoderFactory<ClientRegistration> {
-    public JwtDecoder createDecoder(ClientRegistration reg) {
-      return new CustomJwtDecoder();
-    }
-  }
-}
-
-@Slf4j
-class CustomJwtDecoder implements JwtDecoder {
-
-  public static final ObjectMapper OBJECT_MAPPER = new ObjectMapper();
-
-  @Override
-  public Jwt decode(String token) throws JwtException {
-    JWT jwt;
-    try {
-      jwt = JWTParser.parse(token);
-      return createJwt(token, jwt);
-    } catch (ParseException e) {
-      log.error("Token parsing exception : ", e);
-    }
-    return null;
-  }
-
-  private Jwt createJwt(String token, JWT parsedJwt) {
-    try {
-      Map<String, Object> headers = new LinkedHashMap<>(parsedJwt.getHeader().toJSONObject());
-      Map<String, Object> claimsMap = new HashMap<>();
-      Map<String, Object> claims;
-      if (parsedJwt instanceof SignedJWT) {
-        claims = getJWTClaimsSet(parsedJwt).getClaims();
-      } else claims = parsedJwt.getJWTClaimsSet().getClaims();
-
-      for (String key : claims.keySet()) {
-        Object value = claims.get(key);
-        if (key.equals("exp") || key.equals("iat")) {
-          value = ((Date) value).toInstant();
-        }
-        claimsMap.put(key, value);
-      }
-      return Jwt.withTokenValue(token)
-          .headers(h -> h.putAll(headers))
-          .claims(c -> c.putAll(claimsMap))
-          .build();
-    } catch (Exception ex) {
-      log.error("Exception while creating JWT : ", ex);
-      if (ex.getCause() instanceof ParseException) {
-        throw new JwtException("There is a problem parsing the JWT: " + ex.getMessage());
-      } else {
-        throw new JwtException("There is a problem decoding the JWT: " + ex.getMessage());
-      }
-    }
-  }
-
-  public JWTClaimsSet getJWTClaimsSet(JWT parsedJwt) throws ParseException {
-    Payload payload = new Payload(parsedJwt.getParsedParts()[1]);
-    log.info("Payload before : {}", payload);
-    Map<String, Object> json = toJSONObject(payload);
-    log.info("Payload after : {}", json);
-    if (json == null) {
-      log.error("Payload of JWS object is not a valid JSON object");
-      throw new ParseException("Payload of JWS object is not a valid JSON object", 0);
-    } else {
-      return JWTClaimsSet.parse(json);
-    }
-  }
-
-  public Map<String, Object> toJSONObject(Payload payload) {
-    String payloadStr = payload.toString();
-    if (payloadStr == null) {
-      return null;
-    } else {
-      try {
-        return OBJECT_MAPPER.readValue(payloadStr, new TypeReference<>() {});
-      } catch (JsonProcessingException ex) {
-        log.error("Exception while transforming payload to json object ", ex);
-        return null;
-      }
-    }
+    return new CustomJwtDecoder.CustomJwtDecoderFactory();
   }
 }
